Rename and restructure package
