+++ .gitignore
@@ -235,3 +235,4 @@ dmypy.json
 
 # Cython debug symbols
 cython_debug/
+.aider*
